feat: add a new script to help with duplicate checks
If a duplicate it found (the `duplicate` tag is on the page),
the script colors the entire header. This way you don't have
to scroll to top to notice that the candidate is a duplicate.

Also the script automatically opens the duplicate candidate in
a new tab so that the profile already loads in the background
and you can start checking immediately.